+++ loguno-processor/src/main/java/org/loguno/processor/handlers/HandlersProvider.java
@@ -121,6 +121,7 @@ public final class HandlersProvider {
 				.add(AnnotationHandlerPipedExceptionsCatch.AnnotationHandlerError.class)
 
 				.add(AnnotationHandlerMethodThrows.AnnotationHandlerWholeMethod.class)
+				
 
 		.build();
