+++ loguno-processor/src/main/java/org/loguno/processor/utils/JCTreeUtils.java
@@ -21,10 +21,13 @@ import sun.reflect.annotation.AnnotationParser;
 import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.ExecutableElement;
 import java.lang.annotation.Annotation;
+import java.lang.reflect.Array;
 import java.lang.reflect.Method;
 import java.util.*;
+import java.util.function.Supplier;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static org.loguno.processor.configuration.ConfigurationKeys.CLASS_PATTERN;
@@ -38,7 +41,7 @@ public class JCTreeUtils {
 
     public static final String REPEAT_PATTERN = "\\[(.*?)\\]";
 
-    private static VoidAnnotation VOID_ANN = (VoidAnnotation)AnnotationParser.annotationForMap(VoidAnnotation.class, Collections.emptyMap());
+    private static VoidAnnotation VOID_ANN = (VoidAnnotation) AnnotationParser.annotationForMap(VoidAnnotation.class, Collections.emptyMap());
 
     public <E> void findHandlersAndCall(AnnotationTree annotation, E element, ClassContext classContext) {
 
@@ -49,7 +52,7 @@ public class JCTreeUtils {
 
         if (annClass.isPresent()) {
             Stream<? extends AnnotationHandler<?, E>> handlers = handlersProvider.getHandlersByElementAndAnnotation(annClass.get(), element);
-            Annotation annotationObj = JCTreeUtils.createAnnotationInstance(annotation, annClass.get());
+            Annotation annotationObj = AnnotationUtils.createAnnotationInstance(annotation, annClass.get());
             handlers.forEach(handler -> {
                 handler.process(annotationObj, element, classContext);
             });
@@ -123,101 +126,8 @@ public class JCTreeUtils {
                 .replace(METHOD_PATTERN, context.getMethods().getLast());
     }
 
-    /**
-     * The method creates the 'real' annotation object based on {@link com.sun.source.tree.AnnotationTree}
-     * Now Strings, Strings[] and primitives are supported as annotation member. Classses, enums, annotations are not supported.
-     */
-    @SuppressWarnings("unchecked")
-    public <A extends Annotation> A createAnnotationInstance(AnnotationTree annotation, Class<A> annotationType) {
-
-        Map<String, String[]> customValues = createAnnotationArgsMap(annotation);
-
-        Map<String, Object> values = new LinkedHashMap<>();
-
-        // Extract default values from annotation
-        for (Method method : annotationType.getDeclaredMethods()) {
-            values.put(method.getName(), method.getDefaultValue());
-            Class<?> type = method.getReturnType();
-            if (customValues.containsKey(method.getName())) {
-                Object o = castValue(type, customValues.get(method.getName()));
-                values.put(method.getName(), o);
-            }
-        }
-
-        return (A) AnnotationParser.annotationForMap(annotationType, values);
-    }
-
-    @SneakyThrows
-    private Object castValue(Class<?> clazz, String... value) {
-
-        if (clazz.isPrimitive()) {
-            switch (clazz.getName()) {
-                case "boolean":
-                    return Boolean.valueOf(value[0]);
-                case "int":
-                    return Integer.valueOf(value[0]);
-                case "long":
-                    return Long.valueOf(value[0]);
-                case "double":
-                    return Double.valueOf(value[0]);
-                case "float":
-                    return Float.valueOf(value[0]);
-                case "byte":
-                    return Byte.valueOf(value[0]);
-            }
-        }
-
-        if (clazz.isEnum()) {
-            throw new UnsupportedOperationException();
-        }
-
-        if (clazz.isArray()) {
-            return value;
-        }
-
-        return value[0];
-    }
-
-    private Map<String, String[]> createAnnotationArgsMap(AnnotationTree annotation) {
-        List<? extends ExpressionTree> arguments = annotation.getArguments();
-
-        Map<String, String[]> values = new LinkedHashMap<>();
-
-        for (ExpressionTree argument : arguments) {
-            if (argument instanceof JCTree.JCAssign) {
-                JCTree.JCAssign aArgument = (JCTree.JCAssign) argument;
-                String name = ((JCTree.JCIdent) aArgument.lhs).getName().toString();
-                values.put(name, castElement(aArgument.rhs));
-            } else {
-                values.put("value", castElement(argument));
-            }
-        }
-        return values;
-    }
 
 
-    private String[] castElement(ExpressionTree argument) {
 
-        String[] result = new String[0];
-
-        if (argument instanceof JCTree.JCLiteral) {
-            JCTree.JCLiteral aArgument = (JCTree.JCLiteral) argument;
-            result = new String[]{aArgument.getValue().toString()};
-
-        } else if (argument instanceof JCTree.JCFieldAccess) {
-            JCTree.JCFieldAccess aArgument = (JCTree.JCFieldAccess) argument;
-            result = new String[]{aArgument.toString()};
-
-        } else if (argument instanceof JCTree.JCNewArray) {
-            JCTree.JCNewArray aArgument = (JCTree.JCNewArray) argument;
-
-            result = aArgument.elems.stream()
-                    .map(JCTreeUtils::castElement)
-                    .flatMap(Arrays::stream)
-                    .toArray(String[]::new);
-        }
-
-        return result;
-    }
 
 }
